Apply clippy suggestions

+++ src/buffer.rs
@@ -490,10 +490,10 @@ where P::Subpixel: 'static {
     pub fn new(width: u32, height: u32) -> ImageBuffer<P, Vec<P::Subpixel>> {
         ImageBuffer {
             data: vec![Zero::zero();
-                      (width as u64
-                      * height as u64
-                      * (<P as Pixel>::channel_count() as u64)
-                      ) as usize],
+                      (width as usize
+                      * height as usize
+                      * (<P as Pixel>::channel_count() as usize)
+                      )],
             width: width,
             height: height,
             _phantom: PhantomData,
